refactor 303: 별점 관리를 Integer 로 변경하고 dto에서 값 검증

+++ backend/src/main/java/com/mysite/knitly/domain/product/review/dto/ReviewListResponse.java
@@ -1,14 +1,13 @@
 package com.mysite.knitly.domain.product.review.dto;
 
 import com.mysite.knitly.domain.product.review.entity.Review;
-import com.mysite.knitly.domain.product.review.service.ReviewService;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
 public record ReviewListResponse(
         Long reviewId,
-        Byte rating,
+        Integer rating,
         String content,
         LocalDateTime createdAt,
         String userName,
